Refractor data Handling in GroceryList
